requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
umap-learn>=0.5
matplotlib>=3.7
torch>=2.0
jupyter>=1.0
nbformat>=5.9
